[ANDROID] Vendor Add/Delete Product & Refactoring (#475)
* endpoint functions and layouts added
* fixed unit price bug
* necessary navigations & buttons added
* DELETE /vendor/product added
* visibilities and design updated
* design updates
* notification images updated
* fixed number format bug
* order num added to vendor profile

+++ ecommerce/android/Getflix/app/src/main/java/com/example/getflix/models/VendorOrderModel.kt
@@ -10,8 +10,8 @@ data class VendorOrderModel (
     @SerializedName("amount") val amount : Int,
     @SerializedName("product") val product : ProductModel,
     @SerializedName("status") val status : String,
-    @SerializedName("unit_price") val unitPrice : Int,
-    @SerializedName("total_price") val totalPrice : Int,
+    @SerializedName("unit_price") val unitPrice : Double,
+    @SerializedName("total_price") val totalPrice : Double,
     @SerializedName("purchase_date") val purchaseDate : String,
     @SerializedName("address") val address : AddressModel
 ) :Parcelable
